requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==13.15.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
